requirements.txt
qdrant-client
google-generativeai
python-dotenv
Pillow
